feat: pivot just idk delete all unused

+++ server/api/auth/logout.post.ts
@@ -1,13 +1,12 @@
 export default defineEventHandler(async (event) => {
-  const cookies = parseCookies(event);
-  const refreshToken = cookies.refresh_token;
-
-  if (refreshToken) {
-    await prisma.user.updateMany({
-      where: { refreshToken },
-      data: { refreshToken: null },
-    });
+  if (!event.context.auth) {
+    return jsend.fail(null);
   }
+
+  await prisma.user.updateMany({
+    where: { id: event.context.auth.id },
+    data: { refreshToken: null },
+  });
 
   deleteCookie(event, "auth_token", {
     httpOnly: true,
